feature: classesController create method
